ref(avatars): Move most of UserAvatarPhotoView into a base class
